Added additional options to increase the speed of execution
Added an option to skip the reset, before and after workload execution
as well as skip the fw check after the load stage. These arguments are
no free however, not running reset can cause subsequent workloads
(including burnin) to not load correctly and not running the fw check
may cause hard to diagnose failures if the firmware is corrupted.

Signed-off-by: Daniel Rosen <drosen@tenstorrent.com>