Only use GO_BUSY/GO_IDLE power messages on BH if tt-kmd < v2.6.0
Enforce the use of the set_power_state interface for kmd >= v2.6.0. Ask users to
update FW if the ioctl fails. It can also fail with EINVAL if the chip doesn't
provide a good response to the message for whatever reason, i.e. it could be
hung, so also mention that a reset might be required.

This will change after we add a fix to the Luwen side to handle this error.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>